+++ packages/core/src/odata/common/index.ts
@@ -3,7 +3,7 @@ export * from './edm-types';
 export * from './entity';
 export * from './entity-builder';
 export * from './filter';
-export * from './odate-batch-consts';
+export * from './odata-batch-consts';
 export * from './order';
 export * from './request';
 export * from './request-builder';