+++ packages/core/src/odata/common/odate-batch-consts.ts
@@ -1,4 +1,0 @@
-export const part_content_type_line = 'Content-Type: application/http';
-export const content_transfer_encoding_line =
-  'Content-Transfer-Encoding: binary';
-export const http_version = 'HTTP/1.1';